requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3
APScheduler==3.10.4
